Move withdrawal checks into Cas1UserAccesService
This code is CAS1 specific
