+++ src/test/kotlin/uk/gov/justice/digital/hmpps/approvedpremisesapi/unit/service/UserAccessServiceTest.kt
@@ -16,20 +16,15 @@ import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.ApprovedPremises
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.ApprovedPremisesEntityFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.BookingEntityFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.LocalAuthorityEntityFactory
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.PlacementApplicationEntityFactory
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.PlacementRequestEntityFactory
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.PlacementRequirementsEntityFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.ProbationDeliveryUnitEntityFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.ProbationRegionEntityFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.TemporaryAccommodationApplicationEntityFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.TemporaryAccommodationAssessmentEntityFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.TemporaryAccommodationPremisesEntityFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.UserEntityFactory
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.UserRoleAssignmentEntityFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.factory.cas3.Cas3PremisesEntityFactory
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserRole
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserRole.CAS1_CRU_MEMBER
-import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserRole.CAS1_JANITOR
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserRole.CAS3_REFERRER
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.jpa.entity.UserRole.CAS3_REPORTER
 import uk.gov.justice.digital.hmpps.approvedpremisesapi.service.LaoStrategy
@@ -901,255 +896,6 @@ class UserAccessServiceTest {
     user.addRoleForUnitTest(CAS3_REFERRER)
 
     assertThat(userAccessService.userHasAllRegionsAccess(user, ServiceName.temporaryAccommodation)).isFalse()
-  }
-
-  @Nested
-  inner class UserMayWithdrawApplication {
-
-    @Test
-    fun `userMayWithdrawApplication returns true if application was created by user`() {
-      val application = ApprovedPremisesApplicationEntityFactory()
-        .withCreatedByUser(user)
-        .produce()
-
-      assertThat(userAccessService.userMayWithdrawApplication(user, application)).isTrue
-    }
-
-    @Test
-    fun `userMayWithdrawApplication returns false if application was not created by user`() {
-      val application = ApprovedPremisesApplicationEntityFactory()
-        .withCreatedByUser(anotherUserInRegion)
-        .produce()
-
-      assertThat(userAccessService.userMayWithdrawApplication(user, application)).isFalse()
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = UserRole::class)
-    fun `userMayWithdrawApplication returns true if submitted and has CRU_MEMBER or JANITOR role`(
-      role: UserRole,
-    ) {
-      val otherUser = UserEntityFactory()
-        .withProbationRegion(probationRegion)
-        .produce()
-
-      otherUser.addRoleForUnitTest(role)
-
-      val application = ApprovedPremisesApplicationEntityFactory()
-        .withCreatedByUser(user)
-        .withSubmittedAt(OffsetDateTime.now())
-        .produce()
-
-      val canCancelApplication = listOf(CAS1_CRU_MEMBER, CAS1_JANITOR).contains(role)
-
-      assertThat(userAccessService.userMayWithdrawApplication(otherUser, application)).isEqualTo(canCancelApplication)
-    }
-
-    @Test
-    fun `userMayWithdrawApplication returns false if not submitted and has CAS1_CRU_MEMBER role`() {
-      val cruMember = UserEntityFactory()
-        .withProbationRegion(probationRegion)
-        .produce()
-
-      cruMember.roles.add(
-        UserRoleAssignmentEntityFactory()
-          .withUser(user)
-          .withRole(CAS1_CRU_MEMBER)
-          .produce(),
-      )
-
-      val application = ApprovedPremisesApplicationEntityFactory()
-        .withCreatedByUser(user)
-        .produce()
-
-      assertThat(userAccessService.userMayWithdrawApplication(cruMember, application)).isFalse()
-    }
-
-    @Test
-    fun `userMayWithdrawApplication returns false if not CAS1`() {
-      val application = TemporaryAccommodationApplicationEntityFactory()
-        .withCreatedByUser(anotherUserInRegion)
-        .withProbationRegion(probationRegion)
-        .withSubmittedAt(null)
-        .produce()
-
-      assertThat(userAccessService.userMayWithdrawApplication(user, application)).isFalse
-    }
-  }
-
-  @Nested
-  inner class UserMayWithdrawPlacementRequest {
-
-    @Test
-    fun `userMayWithdrawPlacementRequest returns true if application was created by user`() {
-      val application = ApprovedPremisesApplicationEntityFactory()
-        .withCreatedByUser(user)
-        .produce()
-
-      val assessment = ApprovedPremisesAssessmentEntityFactory()
-        .withApplication(application)
-        .withAllocatedToUser(user)
-        .produce()
-
-      val placementRequest = PlacementRequestEntityFactory()
-        .withApplication(application)
-        .withAssessment(assessment)
-        .withPlacementRequirements(
-          PlacementRequirementsEntityFactory()
-            .withApplication(application)
-            .withAssessment(assessment)
-            .produce(),
-        )
-        .produce()
-
-      assertThat(userAccessService.userMayWithdrawPlacementRequest(user, placementRequest)).isTrue
-    }
-
-    @Test
-    fun `userMayWithdrawPlacementRequest returns false if application was not created by user and doesn't have CAS1_CRU_MEMBER role`() {
-      val application = ApprovedPremisesApplicationEntityFactory()
-        .withCreatedByUser(anotherUserInRegion)
-        .produce()
-
-      val assessment = ApprovedPremisesAssessmentEntityFactory()
-        .withApplication(application)
-        .withAllocatedToUser(user)
-        .produce()
-
-      val placementRequest = PlacementRequestEntityFactory()
-        .withApplication(application)
-        .withAssessment(assessment)
-        .withPlacementRequirements(
-          PlacementRequirementsEntityFactory()
-            .withApplication(application)
-            .withAssessment(assessment)
-            .produce(),
-        )
-        .produce()
-
-      assertThat(userAccessService.userMayWithdrawPlacementRequest(user, placementRequest)).isFalse
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = UserRole::class)
-    fun `userMayWithdrawPlacementRequest returns true if user has CAS1_CRU_MEMBER or JANITOR role`(
-      role: UserRole,
-    ) {
-      val otherUser = UserEntityFactory()
-        .withProbationRegion(probationRegion)
-        .produce()
-
-      otherUser.addRoleForUnitTest(role)
-
-      val application = ApprovedPremisesApplicationEntityFactory()
-        .withCreatedByUser(user)
-        .produce()
-
-      val assessment = ApprovedPremisesAssessmentEntityFactory()
-        .withApplication(application)
-        .withAllocatedToUser(user)
-        .produce()
-
-      val placementRequest = PlacementRequestEntityFactory()
-        .withApplication(application)
-        .withAssessment(assessment)
-        .withPlacementRequirements(
-          PlacementRequirementsEntityFactory()
-            .withApplication(application)
-            .withAssessment(assessment)
-            .produce(),
-        )
-        .produce()
-
-      val canWithdrawPlacementRequest = listOf(CAS1_CRU_MEMBER, CAS1_JANITOR).contains(role)
-
-      assertThat(userAccessService.userMayWithdrawPlacementRequest(otherUser, placementRequest)).isEqualTo(canWithdrawPlacementRequest)
-    }
-  }
-
-  @Nested
-  inner class UserMayWithdrawPlacementApplication {
-
-    @Test
-    fun `userMayWithdrawPlacementApplication returns true if application was created by user`() {
-      val application = ApprovedPremisesApplicationEntityFactory()
-        .withCreatedByUser(user)
-        .produce()
-
-      val placementApplication = PlacementApplicationEntityFactory()
-        .withApplication(application)
-        .withCreatedByUser(user)
-        .produce()
-
-      assertThat(userAccessService.userMayWithdrawPlacementApplication(user, placementApplication)).isTrue
-    }
-
-    @Test
-    fun `userMayWithdrawPlacementApplication returns false if application was not created by user`() {
-      val application = ApprovedPremisesApplicationEntityFactory()
-        .withCreatedByUser(anotherUserInRegion)
-        .produce()
-
-      val placementApplication = PlacementApplicationEntityFactory()
-        .withApplication(application)
-        .withCreatedByUser(anotherUserInRegion)
-        .produce()
-
-      assertThat(userAccessService.userMayWithdrawPlacementApplication(user, placementApplication)).isFalse
-    }
-
-    @ParameterizedTest
-    @EnumSource(value = UserRole::class)
-    fun `userMayWithdrawPlacementApplication returns true if submitted and has CRU_MEMBER or JANITOR role`(
-      role: UserRole,
-    ) {
-      val otherUser = UserEntityFactory()
-        .withProbationRegion(probationRegion)
-        .produce()
-
-      otherUser.addRoleForUnitTest(role)
-
-      val application = ApprovedPremisesApplicationEntityFactory()
-        .withCreatedByUser(user)
-        .withSubmittedAt(OffsetDateTime.now())
-        .produce()
-
-      val placementApplication = PlacementApplicationEntityFactory()
-        .withApplication(application)
-        .withCreatedByUser(user)
-        .withSubmittedAt(OffsetDateTime.now())
-        .produce()
-
-      val canCancelPlacementApp = listOf(CAS1_CRU_MEMBER, CAS1_JANITOR).contains(role)
-
-      assertThat(userAccessService.userMayWithdrawPlacementApplication(otherUser, placementApplication)).isEqualTo(canCancelPlacementApp)
-    }
-
-    @Test
-    fun `userMayWithdrawPlacementApplication returns false if not submitted and has CAS1_CRU_MEMBER role`() {
-      val cruMember = UserEntityFactory()
-        .withProbationRegion(probationRegion)
-        .produce()
-
-      cruMember.roles.add(
-        UserRoleAssignmentEntityFactory()
-          .withUser(user)
-          .withRole(CAS1_CRU_MEMBER)
-          .produce(),
-      )
-
-      val application = ApprovedPremisesApplicationEntityFactory()
-        .withCreatedByUser(anotherUserInRegion)
-        .produce()
-
-      val placementApplication = PlacementApplicationEntityFactory()
-        .withApplication(application)
-        .withCreatedByUser(anotherUserInRegion)
-        .withSubmittedAt(null)
-        .produce()
-
-      assertThat(userAccessService.userMayWithdrawPlacementApplication(user, placementApplication)).isFalse
-    }
   }
 
   @Nested
